Add remote_config python script
This python script is intended to be used for untouched TwinCAT/BSD
systems which should be automatically provisioned via network.
Therefore the python script opens a ssh session, transmits a
config file (should be prepared to your needs) via sftp to the
remote system and afterwards executes this script as root by
using the doas command to gain root rights.